% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmdagg}
\alias{mmdagg}
\title{Find test statistic for MMD test}
\usage{
mmdagg(x, y)
}
\arguments{
\item{x}{A matrix}

\item{y}{A matrix}
}
\value{
A numeric vector with test statistics
}
\description{
Find test statistic for MMD test
}
\keyword{internal}
