% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecheck.R
\name{timecheck}
\alias{timecheck}
\title{test function}
\usage{
timecheck(dta, TS, typeTS, TSextra)
}
\arguments{
\item{dta}{data set}

\item{TS}{test statistics}

\item{typeTS}{format of TS}

\item{TSextra}{additional info TS}
}
\value{
Mean computation time
}
\description{
test function
}
\keyword{internal}
